fix: 64-bit packet counters + make ingest a single lock transaction
Counters: receive/loss/queue_drop were uint32_t -- at Horizon rates they
wrap in ~20 days (single echo), corrupting cumulative loss% and the CSV
snapshots on a 24/7 deployment. Widen to uint64_t end-to-end (struct,
[LivoxStats] printf via PRIu64, snapshot CSV via %llu). Add
publish_packet_count, incremented when a real (non zero-fill) packet is
popped and converted in any of the three publish paths -- this is the
"did the driver actually publish" signal the recv-only stats cannot
provide (a stuck consumer shows recv>0, published=0), groundwork for the
publish-stall watchdog.

Locking: StorageRawPacket checked connect_state and updated statistics
before taking data_lock_, so a packet callback racing a disconnect could
pass the Off check, then re-create the queue ResetLidar had just freed
(stale queue revival) and write fields concurrently with memset. Move
the whole ingest -- Off re-check, metadata/statistics update, queue
alloc/push -- under the per-lidar lock as one transaction. The consumer
already holds the same lock across its poll (DistributeLidarData), so
ordering is unchanged; the added hold time is microseconds of counter
arithmetic.

+++ livox_ros_driver/livox_ros_driver/health_logger.h
@@ -62,8 +62,9 @@ class HealthLogger {
    *  totals without missing anything in between. */
   void LogSnapshot(int handle, const char *bcode, const char *state,
                    const char *temp, const char *fan, const char *motor,
-                   unsigned dirty, const char *system, unsigned recv_total,
-                   unsigned loss_total, unsigned drop_total, double loss_pct,
+                   unsigned dirty, const char *system,
+                   unsigned long long recv_total, unsigned long long loss_total,
+                   unsigned long long drop_total, double loss_pct,
                    unsigned disc) {
     if (!enabled_) {
       return;
@@ -72,7 +73,7 @@ class HealthLogger {
     NowStamp(ts, sizeof(ts));
     char row[384];
     snprintf(row, sizeof(row),
-             "%s,%d,%s,%s,%s,%s,%s,%u,%s,%u,%u,%u,%.3f,%u", ts, handle,
+             "%s,%d,%s,%s,%s,%s,%s,%u,%s,%llu,%llu,%llu,%.3f,%u", ts, handle,
              (bcode && bcode[0]) ? bcode : "?", state, temp, fan, motor, dirty,
              system, recv_total, loss_total, drop_total, loss_pct, disc);
     Write("snapshot",
